Cleanup generators
